requirements.txt
appdirs==1.4.4
backcall==0.2.0
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
flake8==3.8.4
idna==2.10
importlib-metadata==1.7.0
indico-client==4.5.0
ipdb==0.13.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
msgpack==1.0.0
msgpack-numpy==0.4.4.3
mypy-extensions==0.4.3
numpy==1.19.2
pandas==1.1.2
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.9.27
requests==2.24.0
six==1.15.0
toml==0.10.1
traitlets==5.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
zipp==3.3.0
